pyproject.toml
[tool.poetry]
name = "precheck"
version = "0.1.0"
description = "Tool for making pre-check sheet"
authors = ["Kirill"]

[tool.poetry.dependencies]
python = "^3.7"
request = "^2019.4"
google-api-python-client = "^1.7"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.0"
pyopenssl = "^19.0"
gspread = "^3.1"
oauth2client = "^4.1"
beautifulsoup4 = "^4.8"

[tool.poetry.dev-dependencies]
rope = "^0.14.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
